pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "github-todoist-sync"
version = "0.1.0"
description = "A tool to sync GitHub issues and PRs to Todoist tasks."
authors = [{ name = "Remco Sprooten", email = "remco@sprooten.org" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.6"
dependencies = [
    "PyGithub>=1.55",
    "todoist-api-python>=1.3.0",
]

[project.scripts]
github-todoist-sync = "github_todoist_sync.sync:main"
github-todoist-sync-init = "github_todoist_sync.sync:init_config"
